# apps/api/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "AdminPlatformStatsOut": {
        "description": "Global system-wide statistics for superadmins.",
        "properties": {
          "activeRecurring": {
            "title": "Activerecurring",
            "type": "integer"
          },
          "totalAmount": {
            "title": "Totalamount",
            "type": "string"
          },
          "totalDebts": {
            "title": "Totaldebts",
            "type": "integer"
          },
          "totalExpenses": {
            "title": "Totalexpenses",
            "type": "integer"
          },
          "totalUsers": {
            "title": "Totalusers",
            "type": "integer"
          }
        },
        "required": [
          "totalUsers",
          "totalExpenses",
          "totalAmount",
          "totalDebts",
          "activeRecurring"
        ],
        "title": "AdminPlatformStatsOut",
        "type": "object"
      },
...
      "AdminUserListOut": {
        "description": "List of users with pagination total.",
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/AdminUserItemOut"
            },
            "title": "Items",
            "type": "array"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "total"
        ],
        "title": "AdminUserListOut",
        "type": "object"
      },
...
      "AdminUserDetailOut": {
        "description": "Comprehensive user data summary for superadmins.",
        "properties": {
          "budget": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AdminUserBudgetOut"
              },
              {
                "type": "null"
              }
            ],
            "default": null
          },
          "currentMonthExpense": {
            "default": "0.00",
            "title": "Currentmonthexpense",
            "type": "string"
          },
          "netDebt": {
            "default": "0.00",
            "title": "Netdebt",
            "type": "string"
          },
          "totalBorrow": {
            "default": "0.00",
            "title": "Totalborrow",
            "type": "string"
          },
          "totalLend": {
            "default": "0.00",
            "title": "Totallend",
            "type": "string"
          },
          "user": {
            "$ref": "#/components/schemas/AdminUserItemOut"
          }
        },
        "required": [
          "user"
        ],
        "title": "AdminUserDetailOut",
        "type": "object"
      },
      "AdminUserExpensesOut": {
        "description": "List of expenses for a specific user.",
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ExpenseOut"
            },
            "title": "Items",
            "type": "array"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "total"
        ],
        "title": "AdminUserExpensesOut",
        "type": "object"
      },
      "AdminUserDebtsOut": {
        "description": "List of debts for a specific user.",
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/DebtOut"
            },
            "title": "Items",
            "type": "array"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "total"
        ],
        "title": "AdminUserDebtsOut",
        "type": "object"
      },
      "AdminUserRecurringOut": {
        "description": "List of recurring rules for a specific user.",
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/RecurringOut"
            },
            "title": "Items",
            "type": "array"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "total"
        ],
        "title": "AdminUserRecurringOut",
        "type": "object"
      },
      "AdminUserSuspendIn": {
        "description": "Payload to suspend or unsuspend a user.",
        "properties": {
          "suspended": {
            "default": true,
            "title": "Suspended",
            "type": "boolean"
          }
        },
        "title": "AdminUserSuspendIn",
        "type": "object"
      },
      "AdminUserActionOut": {
        "description": "Result of an admin action on a user.",
        "properties": {
          "success": {
            "default": true,
            "title": "Success",
            "type": "boolean"
          },
          "message": {
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "message"
        ],
        "title": "AdminUserActionOut",
        "type": "object"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "Poi Poi Hisab API",
    "version": "0.28.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/admin/stats": {
      "get": {
        "operationId": "get_platform_stats_api_v1_admin_stats_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminPlatformStatsOut"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Return platform-wide summary statistics.",
        "tags": [
          "admin"
        ]
      }
    },
    "/api/v1/admin/users": {
      "get": {
        "operationId": "list_admin_users_api_v1_admin_users_get",
        "parameters": [
          {
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Q"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminUserListOut"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Search and list all users with aggregated financial metrics.",
        "tags": [
          "admin"
        ]
      }
    },
    "/api/v1/admin/users/{user_id}": {
      "get": {
        "operationId": "get_admin_user_detail_api_v1_admin_users__user_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "user_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "User Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminUserDetailOut"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Fetch complete user profile and aggregated metrics.",
        "tags": [
          "admin"
        ]
      },
      "delete": {
        "summary": "Permanently delete a user",
        "operationId": "delete_admin_user",
        "tags": [
          "admin"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "user_id",
            "required": true,
            "schema": {
              "title": "User Id",
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminUserActionOut"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request (e.g. self-delete)",
            "content": {
              "application/json": {
                "schema": {
                  "": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          },
          "403": {
            "description": "Superadmin access required"
          },
          "404": {
            "description": "User not found"
          }
        }
      }
    },
    "/api/v1/admin/users/{user_id}/expenses": {
      "get": {
        "operationId": "get_admin_user_expenses_api_v1_admin_users__user_id__expenses_get",
        "parameters": [
          {
            "in": "path",
            "name": "user_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "User Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminUserExpensesOut"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Return expenses for a specific user.",
        "tags": [
          "admin"
        ]
      }
    },
    "/api/v1/admin/users/{user_id}/debts": {
      "get": {
        "operationId": "get_admin_user_debts_api_v1_admin_users__user_id__debts_get",
        "parameters": [
          {
            "in": "path",
            "name": "user_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "User Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminUserDebtsOut"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Return all debts for a specific user.",
        "tags": [
          "admin"
        ]
      }
    },
    "/api/v1/admin/users/{user_id}/recurring": {
      "get": {
        "operationId": "get_admin_user_recurring_api_v1_admin_users__user_id__recurring_get",
        "parameters": [
          {
            "in": "path",
            "name": "user_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "User Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminUserRecurringOut"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Return recurring expense rules for a specific user.",
        "tags": [
          "admin"
        ]
      }
    },
    "/api/v1/admin/users/{user_id}/suspend": {
      "post": {
        "summary": "Suspend or reactivate a user",
        "operationId": "suspend_admin_user",
        "tags": [
          "admin"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "user_id",
            "required": true,
            "schema": {
              "title": "User Id",
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdminUserSuspendIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminUserActionOut"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request (e.g. self-suspend)",
            "content": {
              "application/json": {
                "schema": {
                  "": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          },
          "403": {
            "description": "Superadmin access required"
          },
          "404": {
            "description": "User not found"
          }
        }
      }
    }
  }
}